ecommerce-app/recommendation_engine/requirements.txt
pandas==2.1.4
numpy==1.26.3
boto3
mysql-connector-python
pyspark==3.5.3
py4j==0.10.9.7